requirements.txt
# Core Dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# File Uploads
python-dotenv==1.0.0

# AI and Text Processing
google-generativeai==0.3.1
python-docx==1.1.0
PyPDF2==3.0.1

# Development
debugpy==1.8.0  # For Python 3.12 debugging

# Production
gunicorn==21.2.0  # Uncomment for production deployment

#database
pymongo[srv]==3.12
# Python version
# Requires Python 3.12 or higher
